+++ plugins/modules/saos10_classifiers.py
@@ -43,6 +43,9 @@ version_added: 2.9
 short_description: Manage classifiers on Ciena SAOS 10 devices
 description: This module provides declarative management of a classifier
 author: Ciena
+requirements:
+  - ncclient (>=v0.6.4)
+  - xmltodict (>=0.12.0)
 options:
   config:
     description: A dictionary of classifier options